Update bricklayers.py
Rewrite code so that for each perimeter loop it first calculates the shortest distance to the external perimeter. Based on this it shifts the perimeter.
Shifts are performed downwards starting, this ensures the nozzle will not hit the object when travelling without z-hop.
Shifts are performed on layers that do not have overhangs and top solid infil, starting from the second layer (future versions will differ). Currently no extrusion nor flow compensation is applied.
main